Create CreateSuiteRequest
